athletix/requirements.txt
certifi==2019.6.16
chardet==3.0.4
cockroach==0.1
Django==2.2.4
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-mathfilters==0.4.0
django-simple-email-confirmation==0.70
djangorestframework==3.10.3
idna==2.8
Pillow==6.1.0
pipenv==2018.11.26
psycopg2==2.8.4
pytz==2019.2
requests==2.22.0
sendmail==2.0
six==1.15.0
sqlparse==0.3.0
tablib==2.0.0
urllib3==1.25.7
virtualenv==16.7.4
virtualenv-clone==0.5.3
